requirements.txt
Django==4.1.2
django-debug-toolbar==3.7.0
django-admin-sortable2==2.1.4
django-tinymce==3.5.0



